move the position of Ipseckey struct
